django_app/requirements.txt
aiokafka==0.12.0
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
beautifulsoup4==4.13.4
billiard==4.2.1
bootstrap4==0.1.0
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.2
decorator==5.2.1
Django==5.2
django-allauth==65.5.0
django-bootstrap4==25.1
django-celery-beat==2.8.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
executing==2.2.0
flower==2.0.1
h11==0.16.0
humanize==4.12.3
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kafka-python==2.2.6
kombu==5.5.3
matplotlib-inline==0.1.7
msgpack==1.1.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
phonenumbers==9.0.4
pillow==11.2.1
pluggy==1.6.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
rabbitmq==0.2.0
redis==5.2.1
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.9.0
yarg==0.1.9
